mnist_digit_classifier/requirements.txt
fastapi
uvicorn
tensorflow
pillow
prometheus_client
